+++ jaziku/core/analysis_interval.py
@@ -26,7 +26,9 @@ from jaziku.utils import  console
 
 
 def get_range_analysis_interval():
-    # range based on analysis interval
+    """
+    Return all start days of analysis interval
+    """
 
     if config_run.settings['analysis_interval'] != "trimester":
         if globals_vars.NUM_DAYS_OF_ANALYSIS_INTERVAL == 5:
@@ -39,6 +41,58 @@ def get_range_analysis_interval():
         return None
 
 
+def get_state_of_data(station):
+    """
+    Calculate and define the state of data based on type of data (daily or monthly)
+    of dependent and independent variable.
+
+    :return:
+        int (1, 2, 3 or 4)
+    """
+    # -------------------------------------------------------------------------
+    # State of the data for process, calculate and write output based on type
+    # of data (daily or monthly) of dependent and independent variable
+    #
+    # | state |  var D  |  var I  |         possible results
+    # |   1   | monthly | monthly |            trimester
+    # |   2   |  daily  | monthly | 5days, 10days, 15days and trimester
+    # |   3   | monthly |  daily  |            trimester
+    # |   4   |  daily  |  daily  | 5days, 10days, 15days and trimester
+    #
+    if station.var_D.frequency_data == "monthly" and station.var_I.frequency_data == "monthly":
+        return 1
+    if station.var_D.frequency_data == "daily" and station.var_I.frequency_data == "monthly":
+        return 2
+    if station.var_D.frequency_data == "monthly" and station.var_I.frequency_data == "daily":
+        return 3
+    if station.var_D.frequency_data == "daily" and station.var_I.frequency_data == "daily":
+        return 4
+
+
+def set_global_state_of_data(stations_list):
+
+    for station in stations_list:
+
+        if globals_vars.STATE_OF_DATA is None:
+            globals_vars.STATE_OF_DATA = get_state_of_data(station)
+            global_frequency_data_of_var_D = station.var_D.frequency_data
+            continue
+        if globals_vars.STATE_OF_DATA != get_state_of_data(station):
+            console.msg_error(_("The station with code '{0}' and name '{1}'\n"
+                                "have data {2} but other stations has data {3}. Jaziku\n"
+                                "requires that all stations have identical frequency data.").format(
+                station.code, station.name, station.var_D.frequency_data, global_frequency_data_of_var_D))
+
+
+def check_analysis_interval():
+
+    if globals_vars.STATE_OF_DATA in [1, 3] and config_run.settings['analysis_interval'] != "trimester":
+        console.msg_error(_("The var_D of stations have data monthly, but you define\n"
+                            "in runfile the analysis interval as '{0}', this must be,\n"
+                            "in this case, as 'trimester' or use data daily.").format(
+            config_run.settings['analysis_interval']))
+
+
 def locate_day_in_analysis_interval(day_for_locate):
     """
     Return corresponding start day of analysis interval where
@@ -49,29 +103,6 @@ def locate_day_in_analysis_interval(day_for_locate):
     for range_item in range_analysis_interval:
         if day_for_locate >= range_item:
             return range_item
-
-
-def check_analysis_interval_and_state_of_data(stations):
-
-    global_state_of_data = None
-
-    for station in stations:
-
-        if global_state_of_data is None:
-            global_state_of_data = station.state_of_data
-            global_frequency_data = station.var_D.frequency_data
-            continue
-        if global_state_of_data != station.state_of_data:
-            console.msg_error(_("The station with code '{0}' and name '{1}'\n"
-                                "have data {2} but other stations has data {3}. Jaziku\n"
-                                "requires that all stations have identical frequency data.").format(
-                station.code, station.name, station.var_D.frequency_data, global_frequency_data))
-
-    if global_state_of_data in [1, 3] and config_run.settings['analysis_interval'] != "trimester":
-        console.msg_error(_("The var_D of stations have data monthly, but you define\n"
-                       "in runfile the analysis interval as '{0}', this must be,\n"
-                       "in this case, as 'trimester' or use data daily.").format(
-            config_run.settings['analysis_interval']))
 
 
 def get_values_in_range_analysis_interval(station, type, year, month, day=None, lag=None):
@@ -137,12 +168,12 @@ def get_values_in_range_analysis_interval(station, type, year, month, day=None, 
                 iter_date += relativedelta(days=1)
 
         if station.var_I.frequency_data == "monthly":
-            if station.state_of_data in [1, 3]:
+            if globals_vars.STATE_OF_DATA in [1, 3]:
                 # get the three values for var_I in this month
                 for iter_month in range(3):
                     var_I_values.append(station.var_I.data[station.var_I.date.index(
                         date(year, month, 1) + relativedelta(months=iter_month - lag))])
-            if station.state_of_data in [2]:
+            if globals_vars.STATE_OF_DATA in [2]:
                 # keep constant value for month
                 if config_run.settings['analysis_interval'] == "trimester":
                     var_I_values.append(station.var_I.data[station.var_I.date.index(