+++ jaziku/modules/forecast/graphs.py
@@ -40,15 +40,15 @@ def forecast_graphs(station):
 
     for lag in config_run.settings['lags']:
 
-        if station.state_of_data in [1, 3]:
-            forecast_month = globals_vars.forecast_date
-            title_date_graphic = _("trim {0} ({1})").format(globals_vars.forecast_date,
+        if globals_vars.STATE_OF_DATA in [1, 3]:
+            forecast_month = config_run.settings['forecast_date']
+            title_date_graphic = _("trim {0} ({1})").format(config_run.settings['forecast_date'],
                 globals_vars.get_trimester_in_text(forecast_month - 1))
             filename_date_graphic = _("trim_{0}").format(forecast_month)
 
-        if station.state_of_data in [2, 4]:
-            forecast_month = globals_vars.forecast_date[0]
-            forecast_day = globals_vars.forecast_date[1]
+        if globals_vars.STATE_OF_DATA in [2, 4]:
+            forecast_month = config_run.settings['forecast_date'][0]
+            forecast_day = config_run.settings['forecast_date'][1]
             title_date_graphic = "{0} {1}".format(globals_vars.get_month_in_text(forecast_month - 1), forecast_day)
             filename_date_graphic = "{0}_{1}".format(globals_vars.get_month_in_text(forecast_month - 1), forecast_day)